Move SoftIRQ implementations to softirq component
