+++ Cargo.toml
@@ -16,6 +16,7 @@ members = [
     "kernel/comps/framebuffer",
     "kernel/comps/input",
     "kernel/comps/network",
+    "kernel/comps/softirq",
     "kernel/comps/time",
     "kernel/comps/virtio",
     "kernel/libs/cpio-decoder",
