+++ Components.toml
@@ -5,6 +5,7 @@ virtio = { name = "aster-virtio" }
 input = { name = "aster-input" }
 block = { name = "aster-block" }
 console = { name = "aster-console" }
+softirq = { name = "aster-softirq" }
 time = { name = "aster-time" }
 framebuffer = { name = "aster-framebuffer" }
 network = { name = "aster-network" }
