+++ kernel/comps/softirq/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "aster-softirq"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ostd = { path = "../../../ostd" }
+component = { path = "../../libs/comp-sys/component" }
+intrusive-collections = "0.9.5"
+spin = "0.9.4"
+
+[features]
