+++ ostd/src/cpu/mod.rs
@@ -13,11 +13,11 @@ cfg_if::cfg_if! {
 }
 
 use bitvec::prelude::BitVec;
-use local::cpu_local_cell;
 use spin::Once;
 
 use crate::{
-    arch::boot::smp::get_num_processors, task::DisabledPreemptGuard, trap::DisabledLocalIrqGuard,
+    arch::boot::smp::get_num_processors, cpu_local_cell, task::DisabledPreemptGuard,
+    trap::DisabledLocalIrqGuard,
 };
 
 /// The number of CPUs.
