+++ ostd/src/trap/mod.rs
@@ -4,10 +4,8 @@
 
 mod handler;
 mod irq;
-pub mod softirq;
 
-pub use handler::in_interrupt_context;
-pub use softirq::SoftIrqLine;
+pub use handler::{in_interrupt_context, register_bottom_half_handler};
 
 pub(crate) use self::handler::call_irq_callback_functions;
 pub use self::irq::{disable_local, DisabledLocalIrqGuard, IrqCallbackFunction, IrqLine};
